refactor(ui): use remote Bluesky avatar URLs for noodles
Noodle avatars only render in the page DOM — the noodle OG image
(Noodle.takumi) doesn't embed them — so the blog's download-and-cache
pipeline is unnecessary here. Resolve each handle to its remote Bluesky
CDN avatar URL at build time (allowed by the CSP img-src), stop writing
public/noodle-avatar, and revert blog.ts to its own avatar helper since
it's no longer shared.
